requirements.txt
boto==2.38.0
django==1.8.8
django-debug-toolbar==1.4
django-flat-theme==1.1.3
django-robots==1.1
django-storages-redux==1.3.1
docutils==0.12
psycopg2==2.6.1
pyinotify==0.9.6
pytz>=2015.7
raven==5.1.1
six==1.10.0
sorl-thumbnail==12.3
sphinx==1.3.4
git+https://github.com/aljosa/django-tinymce@fa2675893637b43f79697ccc853a3a5574da864a
